requirements.txt
aiofiles==23.2.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.36.0
Deprecated==1.2.14
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.3
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
gradio==4.44.1
gradio_client==1.3.0
graphene==3.4
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.17.0
mlflow-skinny==2.17.0
mysql-connector-python==9.1.0
numpy==2.0.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
protobuf==5.28.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
rsa==4.9
ruff==0.7.0
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.41.0
threadpoolctl==3.5.0
tomli==2.0.2
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
waitress==3.0.0
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
